Adding prepend method.
Function prepend cannot take an argument of type T because we violate
the Liskov Substitution Principle (co variant type parameter T is found
as type parameter of the function prepend). We fix that, by using an
upper bound. We pass an argument that is a super type of the list
element type T.