Re-designing Books database as a Set instead of a Seq (List).
The problems fixed with the previous commits are a sign that we started
off with the wrong data structure. In databases, the order in which the
rows appear shouldn't matter. Sets is an abstract data structure that
models this and also have the advantage that duplicates are eliminated
by design. This commit reverts the fixes of the previous two commits.